requirements.txt
numpy
matplotlib
Pillow
scikit-learn